Split amendment content into modules + plugins
Also rewrites the relevant tests to use py.test style
